django-files/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
cfgv==3.4.0
charset-normalizer==3.2.0
code-stego==0.1.0
discord.py==2.3.2
distlib==0.3.7
Django==4.0
djangorestframework==3.14.0
filelock==3.12.3
flake8==4.0.1
flake8-docstrings==1.6.0
frozenlist==1.4.0
gunicorn==21.2.0
identify==2.5.27
idna==3.4
isort==5.10.1
mccabe==0.6.1
multidict==6.0.4
nodeenv==1.8.0
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
platformdirs==3.10.0
pre-commit==2.17.0
pycodestyle==2.8.0
pydocstyle==6.3.0
pyflakes==2.4.0
Pygments==2.16.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
snowballstemmer==2.2.0
sqlparse==0.4.4
toml==0.10.2
tzdata==2023.3
virtualenv==20.24.4
yarl==1.9.2
